drawGrid now uses width and height instead of resolution in fillRect call
drawGrid used to use resolution to determine the size of the squares, but trying to draw more squares posed a problem. If you increased the amount of squares, you also increased their size.

Now width and height are added in the global scope and used by drawGrid().
Allowing for seperate fine tuning of square sizes and grid size